Move Navbar to general components folder for consistency
